+++ src/main/java/com/example/backoffice/domain/question/service/QuestionsServiceV1.java
@@ -23,9 +23,9 @@ public interface QuestionsServiceV1 {
             Long evaluationId, Members loginMember,
             QuestionsRequestDto.DeleteDto requestDto);
 
-    QuestionsResponseDto.SubmitAllDto submitAll(
-            Long evaluationId, Members loginMember,
-            QuestionsRequestDto.SubmitAllDto requestDto);
+    void submitOne(QuestionsRequestDto.SubmitOneDto requestDto, Members loginMember);
+
+    void findByEvaluationIdAndNumber(Long evaluationId, Long number);
 
     Questions findById(Long questionId);
 }
